// design-toolkit/components/.storybook/utils/templates/types.ts
/*
 * Copyright 2025 Hypergiant Galactic Systems Inc. All rights reserved.
 * This file is licensed to you under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License. You may obtain a copy
 * of the License at https://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software distributed under
 * the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTATIONS
 * OF ANY KIND, either express or implied. See the License for the specific language
 * governing permissions and limitations under the License.
 */

/** biome-ignore-all lint/suspicious/noExplicitAny: `any` is necessary in these cases */

export interface StateTemplateProps {
  // biome-ignore lint/style/useNamingConvention: Component is a React component
  Component: React.ComponentType<any>;
  baseProps?: Record<string, any>;
  stateProps?: {
    loading?: Record<string, any>;
    error?: Record<string, any>;
    disabled?: Record<string, any>;
    empty?: Record<string, any>;
  };
}